add-description
This bash script takes a directory of images from photoshow, and copies them into a photoprism directory. The images EXIF tags are modified to add any description added into Photoshow to insure you do not have to re-enter that data. This script is provided with no support and feel free to modify it to your needs. It depends on exiftool that must be manually installed on your machine. The script was written on Redhat 7.3. There are variables you can change for the location of your photoshow, photoprism etc. The script is lightly documented. Hope it helps other.